requirements.txt
ultralytics
opencv-python
Pillow
numpy
accelerate

torch
transformers
sentence-transformers

fastapi
uvicorn
python-multipart